feat: extract modal component
