+++ src/utils/samplePost.ts
@@ -1,0 +1,13 @@
+export const samplePosts = [
+  { id: 1, title: 'Post 1', body: 'Body 1' },
+  { id: 2, title: 'Post 2', body: 'Body 2' },
+  { id: 3, title: 'Post 3', body: 'Body 3' },
+  { id: 4, title: 'Post 4', body: 'Body 4' },
+  { id: 5, title: 'Post 5', body: 'Body 5' },
+  { id: 6, title: 'Post 6', body: 'Body 6' },
+  { id: 7, title: 'Post 7', body: 'Body 7' },
+  { id: 8, title: 'Post 8', body: 'Body 8' },
+  { id: 9, title: 'Post 9', body: 'Body 9' },
+  { id: 10, title: 'Post 10', body: 'Body 10' },
+  { id: 11, title: 'Post 11', body: 'Body 11' },
+];
